+++ EMS/common-bundle/src/Common/Metric/MetricCollector.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace EMS\CommonBundle\Common\Metric;
 
 use EMS\CommonBundle\Common\Cache\Cache;
-use EMS\CommonBundle\Common\Standard\DateTime;
+use EMS\Helpers\Standard\DateTime;
 use Prometheus\CollectorRegistry;
 use Prometheus\MetricFamilySamples;
 use Prometheus\Storage\Adapter;
